+++ test/pay/paddle_billing/customer_test.rb
@@ -8,4 +8,8 @@ class Pay::PaddleBilling::CustomerTest < ActiveSupport::TestCase
   test "paddle cannot create a charge without options" do
     assert_raises(Paddle::Errors::ForbiddenError) { @pay_customer.charge(1000) }
   end
+
+  test "paddle billing subscribe is not supported" do
+    assert_raises(Pay::NotSupportedError) { @pay_customer.subscribe }
+  end
 end